chore: move chains to details dir to keep things tidy
