+++ src/chains/details/arbitrum.ts
@@ -1,6 +1,6 @@
-import { ChainInfo, SupportedChainId } from '../common/types/chains'
-import { baseNativeCurrency } from '../common/consts/tokens'
-import { RAW_CHAINS_FILES_PATH } from './const'
+import { ChainInfo, SupportedChainId } from '../../common/types/chains'
+import { baseNativeCurrency } from '../../common/consts/tokens'
+import { RAW_CHAINS_FILES_PATH } from '../const'
 
 export const arbitrumOneLogoLight = `${RAW_CHAINS_FILES_PATH}/images/arbitrum-one-logo-light.svg`
 export const arbitrumOneLogoDark = `${RAW_CHAINS_FILES_PATH}/images/arbitrum-one-logo-dark.svg`
