+++ src/chains/details/gnosis.ts
@@ -1,6 +1,6 @@
-import { ChainInfo, SupportedChainId } from '../common/types/chains'
-import { baseNativeCurrency, WRAPPED_NATIVE_CURRENCIES } from '../common/consts/tokens'
-import { RAW_CHAINS_FILES_PATH } from './const'
+import { ChainInfo, SupportedChainId } from '../../common/types/chains'
+import { baseNativeCurrency, WRAPPED_NATIVE_CURRENCIES } from '../../common/consts/tokens'
+import { RAW_CHAINS_FILES_PATH } from '../const'
 
 export const gnosisChainLogo = `${RAW_CHAINS_FILES_PATH}/images/gnosis-logo.svg`
